+++ src/__tests__/forLoop.test.js
@@ -2,7 +2,8 @@ import {
   getForLoopSyntax, 
   getSequenceArray, 
   getContentWithReplacedForLoopsSyntax, 
-  getForLoopSyntaxInContent 
+  getForLoopSyntaxInContent, 
+  isStatementObjectValid
 } from '../../src/forLoopHelpers'
 
 jest.mock('faker', () => ({
@@ -31,6 +32,13 @@ describe('Helper\'s functions to process for loops', () => {
       it('should give array from array notation', () => {
         const arrayNotation = '[1, 2]'
         const expectedResult = [1, 2]
+
+        expect(getSequenceArray(arrayNotation)).toStrictEqual(expectedResult)
+      })
+
+      it('should give array of objects from array notation', () => {
+        const arrayNotation = "[{'a': '1', 'b': 2}, {\"c\":\"3\", 'd':4}]"
+        const expectedResult = [{ 'a': '1', 'b': 2 }, { 'c': '3', 'd': 4 }]
 
         expect(getSequenceArray(arrayNotation)).toStrictEqual(expectedResult)
       })
@@ -101,11 +109,37 @@ describe('Helper\'s functions to process for loops', () => {
     expect(getContentWithReplacedForLoopsSyntax(content)).toBe(expectedResult)
   })
 
+  it('should return the proceeded for loop with objects in statement', () => {
+    const content = `["#for i in [{'t':'snake'},{'t':'mouse'}]",{"type": "[i.t]"},"#endfor"]`
+    const expectedResult = '[{"type": "snake"},{"type": "mouse"}]'
+
+    expect(getContentWithReplacedForLoopsSyntax(content)).toBe(expectedResult)
+  })
+  
+  it('should return the proceeded for loop with objects in statement and cast x to number notation', () => {
+    const content = `["#for i in [{'t':'1'},{'t':'2'}]",{"type": "n:[i.t]"},"#endfor"]`
+    const expectedResult = '[{"type": 1},{"type": 2}]'
+
+    expect(getContentWithReplacedForLoopsSyntax(content)).toBe(expectedResult)
+  })
+
   it('should return proceed nested for loop', () => {
     const content = '["#for i in [1, 2]",["#for ii in [1, 2]",{"parent":"n:[i]","id":"n:[ii]"},"#endfor"],"#endfor"]'
     const expectedResult = 
       '[[{"parent":1,"id":1},{"parent":1,"id":2}],[{"parent":2,"id":1},{"parent":2,"id":2}]]'
     
     expect(getContentWithReplacedForLoopsSyntax(content)).toBe(expectedResult)
+  })
+
+  describe('Statement object validity', () => {
+    it('should detect a valid object', () => {
+      expect(isStatementObjectValid({foo: 'bar'})).toBe(true)
+    })
+    it('should detect an invalid object that use array as value', () => {
+      expect(isStatementObjectValid({foo: []})).toBe(false)
+    })
+    it('should detect an invalid object that use object as value', () => {
+      expect(isStatementObjectValid({foo: {}})).toBe(false)
+    })
   })
 })