+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### ✨ Added
+- support object notation in for-loop array sequence
 
 ## [2.2.0] - 2021-07-30
 ### ✨ Added